Refactor directory structure for tests
